pyproject.toml
[project]
name = "pg2mcp"
version = "0.1.0"
description = "PostgreSQL to MCP Bridge - Expose PostgreSQL objects as MCP resources and tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "fastmcp>=2.10.4",
    "httpx>=0.28.1",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "uvicorn[standard]>=0.30.0",
    "click>=8.0.0",
    "structlog>=25.4.0",
]
